+++ spandex-common/src/test/scala/com/socrata/spandex/common/client/TestESClient.scala
@@ -3,6 +3,7 @@ package com.socrata.spandex.common.client
 import java.nio.file.Files
 
 import com.socrata.spandex.common.{ElasticSearchConfig, SpandexBootstrap}
+import com.socrata.spandex.common.client.ResponseExtensions._
 import org.apache.commons.io.FileUtils
 import org.elasticsearch.client.{Client, Requests}
 import org.elasticsearch.common.settings.ImmutableSettings
@@ -22,16 +23,6 @@ class TestESClient(config: ElasticSearchConfig, local: Boolean = true) extends S
 
   SpandexBootstrap.ensureIndex(config, this)
 
-  def deleteIndex(): Unit = {
-    client.admin().indices().delete(Requests.deleteIndexRequest(config.index))
-  }
-
-  def deleteAllDatasetCopies(): Unit = {
-    client.prepareDeleteByQuery(config.index)
-          .setQuery(termQuery("_type", config.datasetCopyMapping))
-          .execute().actionGet()
-  }
-
   override def close(): Unit = {
     deleteIndex()
     node.close()
@@ -43,5 +34,67 @@ class TestESClient(config: ElasticSearchConfig, local: Boolean = true) extends S
     }
 
     super.close()
+  }
+
+  def deleteIndex(): Unit = {
+    client.admin().indices().delete(Requests.deleteIndexRequest(config.index))
+  }
+
+  def deleteAllDatasetCopies(): Unit = {
+    client.prepareDeleteByQuery(config.index)
+          .setQuery(termQuery("_type", config.datasetCopyMapping))
+          .execute().actionGet()
+  }
+
+  def searchColumnMapsByDataset(datasetId: String): SearchResults[ColumnMap] =
+    client.prepareSearch(config.index)
+          .setTypes(config.columnMapMapping.mappingType)
+          .setQuery(byDatasetIdQuery(datasetId))
+          .execute.actionGet.results[ColumnMap]
+
+  def searchColumnMapsByCopyNumber(datasetId: String, copyNumber: Long): SearchResults[ColumnMap] =
+    client.prepareSearch(config.index)
+          .setTypes(config.columnMapMapping.mappingType)
+          .setQuery(byCopyNumberQuery(datasetId, copyNumber))
+          .execute.actionGet.results[ColumnMap]
+
+  def searchFieldValuesByDataset(datasetId: String): SearchResults[FieldValue] = {
+    val response = client.prepareSearch(config.index)
+                         .setTypes(config.fieldValueMapping.mappingType)
+                         .setQuery(byDatasetIdQuery(datasetId))
+                         .execute.actionGet
+    response.results[FieldValue]
+  }
+
+  def searchFieldValuesByCopyNumber(datasetId: String, copyNumber: Long): SearchResults[FieldValue] = {
+    val response = client.prepareSearch(config.index)
+                         .setTypes(config.fieldValueMapping.mappingType)
+                         .setQuery(byCopyNumberQuery(datasetId, copyNumber))
+                         .execute.actionGet
+    response.results[FieldValue]
+  }
+
+  def searchFieldValuesByColumnId(datasetId: String, copyNumber: Long, columnId: Long): SearchResults[FieldValue] = {
+    val response = client.prepareSearch(config.index)
+                         .setTypes(config.fieldValueMapping.mappingType)
+                         .setQuery(byColumnIdQuery(datasetId, copyNumber, columnId))
+                         .execute.actionGet
+    response.results[FieldValue]
+  }
+
+  def searchFieldValuesByRowId(datasetId: String, copyNumber: Long, rowId: Long): SearchResults[FieldValue] = {
+    val response = client.prepareSearch(config.index)
+                         .setTypes(config.fieldValueMapping.mappingType)
+                         .setQuery(byRowIdQuery(datasetId, copyNumber, rowId))
+                         .execute.actionGet
+    response.results[FieldValue]
+  }
+
+  def searchCopiesByDataset(datasetId: String): SearchResults[DatasetCopy] = {
+    val response = client.prepareSearch(config.index)
+                         .setTypes(config.datasetCopyMapping.mappingType)
+                         .setQuery(byDatasetIdQuery(datasetId))
+                         .execute.actionGet
+    response.results[DatasetCopy]
   }
 }
